LoadDateFromLocalStorage

+++ app.js
@@ -5,10 +5,12 @@ const taskinput = document.querySelector("#task");
 const tasklist = document.querySelector(".collection");
 const clearbutton = document.querySelector(".clear-tasks");
 const filter = document.querySelector("#filter");
-let tasks;
+
 loadEventListener();
 
 function loadEventListener(){
+    //get tasks from localstorage
+    document.addEventListener('DOMContentLoaded',getTasks);
 
     form.addEventListener('submit',addTask);
 
@@ -17,16 +19,50 @@ function loadEventListener(){
     clearbutton.addEventListener('click',clearTask)
 }
 
-function clearTask(e){
-    let li = document.getElementById('li');
-    console.log(tasklist.children);
-    for (i = 1; i <= tasks.length; i++) {
-        console.log(tasklist.children[i]);
-        console.log(i);
-        tasklist.children[0].remove();
+function getTasks(){
+    let tasks;
+    if(localStorage.getItem('tasks') === null){
+        tasks = [];
     }
-    
-    localStorage.clear();
+    else{
+        tasks = JSON.parse(localStorage.getItem('tasks'));
+    }
+    tasks.forEach(function(task){
+        const li = document.createElement('li');
+        li.className = 'collection-item';
+        li.appendChild(document.createTextNode(task));
+        const deleteLink = document.createElement('a');
+        deleteLink.className = 'delete-item secondary-content';
+        deleteLink.innerHTML = '<i class = "fa fa-remove">';
+        li.appendChild(deleteLink);
+        tasklist.appendChild(li);
+    })
+}
+
+function clearTask(e){
+    let tasks;
+    if(localStorage.getItem('tasks') === null){
+        tasks = [];
+    }
+    else{
+        tasks = JSON.parse(localStorage.getItem('tasks'));
+        let li = document.getElementById('li');
+        console.log(tasklist.children);
+        for (i = 1; i <= tasks.length; i++) {
+            try{
+                console.log(tasklist.children[i]);
+                console.log(i);
+                tasklist.children[0].remove();
+            }
+            catch{
+
+            }
+        }
+        
+        localStorage.clear();
+    }
+    console.log(tasks);
+
 }
 
 
@@ -47,7 +83,7 @@ function addTask(e){
 }
 
 function storeTaskInLocalStorage(task){
-
+    let tasks;
     if(localStorage.getItem('tasks') === null){
         tasks = [];
     }
